Separate test fake dataset code to util module * improve comments for dummy dose grid
